requirements.txt
aiofile==3.8.5
aiogram==2.25.1
aiohttp==3.8.4
aiohttpx==0.0.9
aiosignal==1.3.1
alembic==1.10.4
anyio==3.6.2
async-lru==2.0.2
async-openai==0.0.19
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
backoff==2.2.1
beautifulsoup4==4.12.2
caio==0.9.12
certifi==2022.12.7
charset-normalizer==3.1.0
coverage==7.2.4
dill==0.3.6
exceptiongroup==1.1.1
execnet==1.9.0
file-io==0.4.6
frozendict==2.3.7
frozenlist==1.3.3
fsspec==2023.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
lazyops==0.2.39
loguru==0.7.0
lxml==4.9.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
mirakuru==2.5.1
multidict==6.0.4
mypy==1.2.0
mypy-extensions==1.0.0
openai==0.27.4
packaging==23.1
pluggy==1.0.0
port-for==0.6.3
psutil==5.9.5
psycopg==3.1.8
psycopg2-binary==2.9.6
pydantic==1.10.7
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0
pytest-postgresql==4.1.1
pytest-xdist==3.2.1
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
regex==2023.3.23
requests==2.28.2
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.11
SQLAlchemy-Utils==0.41.1
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.9.1
